[all] add model module
