+++ core/data/src/main/java/com/keelim/data/db/AppDatabase.kt
@@ -2,16 +2,18 @@ package com.keelim.data.db
 
 import androidx.room.AutoMigration
 import androidx.room.Database
+import androidx.room.DeleteColumn
+import androidx.room.DeleteTable
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverter
 import androidx.room.TypeConverters
+import androidx.room.migration.AutoMigrationSpec
 import com.keelim.data.db.dao.CnuHistoryDao
 import com.keelim.data.db.dao.CommentDao
 import com.keelim.data.db.dao.HistoryDao
 import com.keelim.data.db.dao.NandaDao
 import com.keelim.data.db.dao.NetworkCacheDao
 import com.keelim.data.db.dao.NoteDao
-import com.keelim.data.db.dao.StationDao
 import com.keelim.data.db.dao.TaskDao
 import com.keelim.data.db.dao.TimerHistoryDao
 import com.keelim.data.model.cache.NetworkCache
@@ -19,9 +21,6 @@ import com.keelim.data.model.entity.CnuHistory
 import com.keelim.data.model.entity.Comment
 import com.keelim.data.model.entity.NandaEntity
 import com.keelim.data.model.entity.Notices
-import com.keelim.data.model.entity.StationEntity
-import com.keelim.data.model.entity.StationSubwayCrossRefEntity
-import com.keelim.data.model.entity.SubwayEntity
 import com.keelim.data.source.local.History
 import com.keelim.data.source.local.LocalTask
 import com.keelim.data.source.local.SimpleHistory
@@ -72,19 +71,24 @@ abstract class NandaAppDatabase : RoomDatabase() {
     abstract fun dataDao(): NandaDao
 }
 
+
+@DeleteTable(tableName = "StationEntity")
+@DeleteTable(tableName = "SubwayEntity")
+@DeleteTable(tableName = "StationSubwayCrossRefEntity")
+class CnuAppDatabaseVersion2To3 : AutoMigrationSpec
+
 @Database(
     entities = [
-        StationEntity::class,
-        SubwayEntity::class,
-        StationSubwayCrossRefEntity::class,
         Comment::class,
         CnuHistory::class,
     ],
-    version = 2,
-    exportSchema = false,
+    version = 3,
+    exportSchema = true,
+    autoMigrations = [
+        AutoMigration(from = 2, to = 3, spec = CnuAppDatabaseVersion2To3::class),
+    ],
 )
 abstract class CnuAppDatabase : RoomDatabase() {
-    abstract fun daoStation(): StationDao
     abstract fun daoComment(): CommentDao
     abstract fun daoHistory(): CnuHistoryDao
 }
