--- core/data/src/main/java/com/keelim/data/db/dao/StationDao.kt
@@ -1,62 +1,0 @@
-/*
- * Designed and developed by 2021 keelim (Jaehyun Kim)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.keelim.data.db.dao
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import androidx.room.Transaction
-import androidx.room.Update
-import com.keelim.data.model.entity.StationEntity
-import com.keelim.data.model.entity.StationSubwayCrossRefEntity
-import com.keelim.data.model.entity.StationWithSubwaysEntity
-import com.keelim.data.model.entity.SubwayEntity
-import kotlinx.coroutines.flow.Flow
-
-@Dao
-interface StationDao {
-
-    @Transaction
-    @Query("SELECT * FROM StationEntity")
-    fun getStationWithSubways(): Flow<List<StationWithSubwaysEntity>>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertStations(station: List<StationEntity>)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertSubways(subways: List<SubwayEntity>)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertCrossReferences(reference: List<StationSubwayCrossRefEntity>)
-
-    @Transaction
-    suspend fun insertStationSubways(stationSubways: List<Pair<StationEntity, SubwayEntity>>) {
-        insertStations(stationSubways.map { it.first })
-        insertSubways(stationSubways.map { it.second })
-        insertCrossReferences(
-            stationSubways.map { (station, subway) ->
-                StationSubwayCrossRefEntity(
-                    station.stationName,
-                    subway.subwayId,
-                )
-            },
-        )
-    }
-
-    @Update
-    suspend fun updateStation(station: StationEntity)
-}
