--- core/data/src/main/java/com/keelim/data/model/Data.kt
@@ -1,41 +1,0 @@
-/*
- * Designed and developed by 2021 keelim (Jaehyun Kim)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.keelim.data.model
-
-import android.os.Parcelable
-import com.google.firebase.firestore.DocumentId
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Data(
-    @DocumentId
-    val id: String? = null,
-
-    @field:JvmField
-    val isFeatured: Boolean? = null,
-
-    val title: String? = null,
-    val actors: String? = null,
-    val country: String? = null,
-    val director: String? = null,
-    val genre: String? = null,
-    val posterUrl: String? = null,
-    val rating: String? = null,
-    val averageScore: Float? = null,
-    val numberOfScore: Int? = null,
-    val releaseYear: Int? = null,
-    val runtime: Int? = null,
-) : Parcelable
