--- core/data/src/main/java/com/keelim/data/model/ErrorMessage.kt
@@ -1,33 +1,0 @@
-/*
- * Designed and developed by 2021 keelim (Jaehyun Kim)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.keelim.data.model
-
-import kotlinx.serialization.SerialName
-
-data class ErrorMessage(
-    @SerialName("code")
-    val code: String? = null,
-    @SerialName("developerMessage")
-    val developerMessage: String? = null,
-    @SerialName("link")
-    val link: String? = null,
-    @SerialName("message")
-    val message: String? = null,
-    @SerialName("status")
-    val status: Int? = null,
-    @SerialName("total")
-    val total: Int? = null,
-)
