--- core/data/src/main/java/com/keelim/data/model/Event.kt
@@ -1,34 +1,0 @@
-/*
- * Designed and developed by 2021 keelim (Jaehyun Kim)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.keelim.data.model
-
-open class Event<out T>(private val content: T) {
-    private var hasBeenHandled = false
-
-    fun getContentIfNotHandled(): T? {
-        return if (hasBeenHandled) { // 이벤트가 이미 처리 되었다면
-            null // null을 반환하고,
-        } else { // 그렇지 않다면
-            hasBeenHandled = true // 이벤트가 처리되었다고 표시한 후에
-            content // 값을 반환합니다.
-        }
-    }
-
-    /**
-     * 이벤트의 처리 여부에 상관 없이 값을 반환합니다.
-     */
-    fun peekContent(): T = content
-}
