+++ core/data/src/main/java/com/keelim/data/model/NandaResponse.kt
@@ -16,32 +16,16 @@
 package com.keelim.data.model
 
 import com.keelim.data.model.entity.NandaEntity2
+import com.keelim.model.NandaResponse
 import java.util.*
 
-data class NandaResponse(
-    val id: Int,
-    val reason: String,
-    val diagnosis: String,
-    val domain_name: String,
-    val class_name: String,
-    val category: Int,
-    val favorite: Int,
-    val createdAt: Long,
-) {
-    fun toEntity(): NandaEntity2 =
-        NandaEntity2(
-            nanda_id = id,
-            reason = reason,
-            diagnosis = diagnosis,
-            class_name = class_name,
-            domain_name = domain_name,
-            category = category,
-            favorite = favorite,
-            created_at = Date(createdAt),
-        )
-}
-
-data class NandasResponse(
-    val items: List<NandaResponse>,
-    val count: Int,
+fun NandaResponse.toEntity() = NandaEntity2(
+    nanda_id = id,
+    reason = reason,
+    diagnosis = diagnosis,
+    class_name = class_name,
+    domain_name = domain_name,
+    category = category,
+    favorite = favorite,
+    created_at = Date(createdAt),
 )
