--- core/data/src/main/java/com/keelim/data/model/entity/StationEntity.kt
@@ -1,25 +1,0 @@
-/*
- * Designed and developed by 2021 keelim (Jaehyun Kim)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.keelim.data.model.entity
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity
-data class StationEntity(
-    @PrimaryKey val stationName: String,
-    val isFavorited: Boolean = false,
-)
