--- core/data/src/main/java/com/keelim/data/model/entity/StationWithSubwaysEntity.kt
@@ -1,30 +1,0 @@
-/*
- * Designed and developed by 2021 keelim (Jaehyun Kim)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.keelim.data.model.entity
-
-import androidx.room.Embedded
-import androidx.room.Junction
-import androidx.room.Relation
-
-data class StationWithSubwaysEntity(
-    @Embedded val station: StationEntity,
-    @Relation(
-        parentColumn = "stationName",
-        entityColumn = "subwayId",
-        associateBy = Junction(StationSubwayCrossRefEntity::class),
-    )
-    val subways: List<SubwayEntity>,
-)
